NotificationUsageDetailScreenUiState をファイル分割
UiStateの定義とScreenのComposableが同じファイルにあったため、
NotificationUsageDetailScreenUiState.kt に分離した
